requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.16.2
arabic-reshaper==3.0.0
attrs==25.3.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cssselect2==0.8.0
defusedxml==0.7.1
dnspython==2.7.0
email-validator==2.1.0
et_xmlfile==2.0.0
Flask==3.0.0
Flask-Login==0.6.3
Flask-CORS==5.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.58.4
fpdf2==2.8.3
frozenlist==1.7.0
greenlet==3.2.3
hijri-converter==2.3.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==2.1.3
multidict==6.4.4
numpy==1.25.2
openpyxl==3.1.2
pandas==2.1.3
Pillow==10.1.0
propcache==0.3.2
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
PyMySQL==1.1.1
pyphen==0.17.2
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2025.2
qrcode==8.2
reportlab==4.0.7
sendgrid==6.10.0
six==1.17.0
SQLAlchemy==2.0.23
starkbank-ecdsa==2.2.0
tinycss2==1.4.0
tinyhtml5==2.0.0
twilio==8.10.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.0.1
WTForms==3.1.1
xlrd==2.0.1
XlsxWriter==3.1.9
yarl==1.20.1
zopfli==0.2.3.post1
cairocffi
gunicorn
requests
resend==0.6.0
openai
pillow-heif
replit-object-storage
apscheduler
google-api-python-client
lxml
